Create smallbets-v1.xml
This is the first mission file created for the smallbets project. It's intended purpose is to model two uuvs facing one another in an underwater environment with a simple plume between them. The environment is created by TerrainGenerator. The env variables are published to scrimmage globally and a sensor plugin subscribes, producing a txt or .csv file for reading-in from BELLHOP