Update OceanParameters.cpp
first guess at including the sub/pub information into the OP.cpp. I based most of my educated guessing on the Boundary example given in the scrimmage documentation (publisher/subscriber section) and the TerrainGenerator plugin (checking and comparing all .h & .cpp files for placement)
I also looked up as much as I could for the c++ code itself, but would like to know what the data structure for the Terrain information is so that I can start reducing it with the proper equations.