Create plume_propagation_ex.m
This file is matlab-based. To use it you need the other files in this folder. Feel free to contact me if you have questions about those files or how this file outputs an ssp profile. 
What it is meant to do is print out a "hot blob" plume based on the parameters passed to it in iterations of the scrimmage run in which we pull out a 2d slice of the underwater environment. 
This 2d slice is what gets modified by this code to artificially input a parameter of interest for reduction by bellhop.